+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [#103] Adds instant validation feedback on input fields.
 - [#15] Adds a button to the header linking to a faucet on testnets.
 - [#11] Adds a token mint button for testnets
+- [#111] Adds payment identifier field to payment dialog.
 
 ## [0.8.0] - 2019-01-25
 ### Changed
@@ -52,6 +53,7 @@ token network.
 [0.7.0]: https://github.com/raiden-network/webui/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/raiden-network/webui/releases/tag/v0.6.0
 
+[#111]: https://github.com/raiden-network/webui/issues/111
 [#105]: https://github.com/raiden-network/webui/issues/105
 [#103]: https://github.com/raiden-network/webui/issues/103
 [#101]: https://github.com/raiden-network/webui/issues/101